% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_cohens_h.R
\name{ci_cohens_h}
\alias{ci_cohens_h}
\title{Confidence Interval for Cohen's h}
\usage{
ci_cohens_h(p1, p2, n1, n2, conf.level = 0.95)
}
\arguments{
\item{p1}{Proportion for Sample 1}

\item{n1}{Proportion for Sample 2}

\item{n2}{Sample size for Sample 2}

\item{conf.level}{Confidence level, default is 0.95}
}
\value{
Lower and upper limits on confidence interval.
}
\description{
Confidence Interval for Cohen's h
}
\examples{
ci_cohens_h(p1 = 0.174, n1 = 27, p2 = .111, n2 = 25, conf.level = 0.95)

}
