% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_risk_ratio.R
\name{ci_risk_ratio}
\alias{ci_risk_ratio}
\title{Confidence Interval for Risk Ratio}
\usage{
ci_risk_ratio(p1, n1, p2, n2, conf.level = 0.95)
}
\arguments{
\item{p1}{Proportion for Sample 1}

\item{n1}{Sample size for Sample 1}

\item{p2}{Proportion for Sample 2}

\item{n2}{Sample size for Sample 2}

\item{conf.level}{Confidence level, default is 0.95}
}
\value{
Lower and upper limits on confidence interval.
}
\description{
Confidence Interval for Risk Ratio
}
\examples{
ci_risk_ratio(p1 = 0.174, n1 = 27, p2 = .111, n2 = 25, conf.level = 0.95)

}
