% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_plot.R
\name{coef_plot}
\alias{coef_plot}
\title{Visually Weighted Coefficient Plot}
\usage{
coef_plot(x, labels = names(coef(x))[start:k], intercept = FALSE, ...)
}
\arguments{
\item{x}{lm object}

\item{labels}{Vector of coefficient labels}

\item{intercept}{Logical indicating whether the intercept should be displayed}
}
\value{
A ggplot object giving the coefficient plot
}
\description{
This function creates a visually weighted coefficient plot
}
