% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{education}
\alias{education}
\title{State Level Education Data}
\format{A data frame with six variables:
\describe{
  \item{\code{state}}{State name}
  \item{\code{postal}}{Two-letter state postal code}
  \item{\code{region}}{Region of the United States (Midwest, Northeast, South, West)}
  \item{\code{salary}}{Average teacher salary in the state}
  \item{\code{sat}}{Average SAT score in the state}
  \item{\code{participation}}{Percentage of students in the state who took the SAT}
}}
\usage{
data(education)
}
\description{
The data includes state-level aggregate data on six attributes. The attributes were collected for
all 50 states in the United STates and the District of Columbia
}
\examples{
data(education)
summary(education)
}
\keyword{datasets}
