% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_density_confidence.R,
%   R/stat_density_watercolor.R
\name{stat_density_confidence}
\alias{stat_density_confidence}
\alias{stat_density_watercolor}
\title{Stat Method for Density with Confidence Envelope}
\arguments{
\item{h}{A normal kernel function is used and `h` is its standard deviation. If this parameter is
omitted, a normal optimal smoothing parameter is used.}

\item{fill}{Fill color for the confidence envelope. The default is `color="skyblue"`}

\item{model}{The model to draw the confidence envelope for. The default is `model="none"` which creates the confidence envelope
from the data. Using `model="normal"` creates the confidence envelope based on a normal distribution.}

\item{k}{Number of bootstrapped smoothers. The default is 1000.}

\item{alpha}{Transparency level for the paths that make up the bootstrapped fitted lines. This may
need to be adjusted if the argument `k=` is changed. The default value is `alpha=0.03`.}

\item{color}{Color for the bootstrapped fitted lines that make up the confidence envelope. The
default is `color="#1D91C0"`}
}
\description{
This function creates a confidence enevelope for the empirical density. The idea came from
functions written by Bowman and Azzalini.

This function creates a visually regression smoother, also called a watercolor plot. The initial idea
came from Solomon Hsiang.
}
\section{Aesthetics}{

These stat uses  `geom_ribbon()` so support the
same aesthetics: `alpha`, `colour`, `linetype` and
`size`.


These stat uses  `geom_line()` so support the
same aesthetics: `alpha`, `colour`, `linetype` and
`size`.
}

