% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_t_dist.R
\name{plot_t_dist}
\alias{plot_t_dist}
\title{Plot t-Distribution and Shaded p-Value}
\usage{
plot_t_dist(x, shade_p_value = TRUE)
}
\arguments{
\item{x}{t.test() or t_test() object}

\item{shade_p_value}{Logical value indicating whether to shade the area under the curve
corresponding to the p-value. Default is TRUE.}
}
\value{
Plot of the t-distribution. If shade_p_value=TRUE, the area under the curve
corresponding to the p-value will also be shaded.
}
\description{
Plot t-Distribution and Shaded p-Value
}
\examples{
my_t = t.test(cars$speed, mu = 16, alternative = "less")
plot_t_dist(my_t)

}
