% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_plots.R
\name{residual_plots}
\alias{residual_plots}
\title{Residual Plots for Linear Models}
\usage{
residual_plots(model, type = "b")
}
\arguments{
\item{model}{Logical indicating whether the model-level partitioning (TRUE) or term-level partitioning (FALSE; default) should be used}

\item{type}{Character string indicating the type of residual plot to return. The default is \code{type="b"} (both). You can also
use \code{type="d"} (density) or \code{type="s"} (scatterplot).}
}
\value{
One or more plots of the standardized residuals. In the density plot, a confidence envelope of reasonable values
based on the normality assumption being true is produced along with the empirical density of the residuals.
In the scatterplot of the standardized residuals versus the fitted values, a confidence envelope of reasonable values
based on the mean residual = 0 (linearity) is produced along with the loess smoother based on the empirical data.
}
\description{
Residual Plots for Linear Models
}
\examples{
lm.1 = lm(mpg ~ 1 + hp + wt, data = mtcars)
residual_plots(lm.1, type = "b")

}
