% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_density.R
\name{sm_density}
\alias{sm_density}
\title{Density plot with Bootstrapped Confidence Envelope}
\usage{
sm_density(
  data,
  x,
  k = 1000,
  boot_envelope = TRUE,
  model = "none",
  rug_color = "black",
  envelope_fill = "skyblue",
  ...
)
}
\arguments{
\item{data}{Dataframe}

\item{x}{Column name (unquoted)}

\item{k}{Number of bootstrap samples}

\item{boot_envelope}{Logical value to indicate whether the bootstrap confidence envelope should be displayed}

\item{model}{Character string indicating the type of bootstrapped confidence envelope to display;}

\item{rug_color}{Color for the rug plot}

\item{envelope_fill}{Fill color for the bootstrapped confidence envelope}
}
\value{
A ggplot object giving the density plot
}
\description{
This function creates a visually weighted coefficient plot
}
