% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_density_confidence.R
\name{stat_density_confidence}
\alias{stat_density_confidence}
\title{Normal Theory Based Confidence Envelope for Density}
\usage{
stat_density_confidence(
  mapping = NULL,
  data = NULL,
  geom = "ribbon",
  position = "identity",
  na.rm = TRUE,
  inherit.aes = TRUE,
  h = NULL,
  fill = "skyblue",
  fade = FALSE,
  model = "none",
  ...
)
}
\arguments{
\item{h}{A normal kernel function is used and \code{h} is its standard deviation. If this parameter is
omitted, a normal optimal smoothing parameter is used.}

\item{fill}{Fill color for the confidence envelope. The default is \code{fill="skyblue"}}

\item{model}{The model to draw the confidence envelope for. The default is \code{model="none"} which creates the
confidence envelope from the data. Using \code{model="normal"} creates the confidence envelope based on a normal distribution.}
}
\description{
This function creates a normal theory based confidence envelope for the empirical density. The idea came from
functions written by Bowman and Azzalini.
}
\section{Aesthetics}{

These stat uses  \code{geom_ribbon()} so support the same aesthetics: \code{alpha}, \code{colour}, \code{fill},
\code{linetype}, \code{group}, and \code{size}.
}

