% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_density_watercolor.R
\name{stat_density_watercolor}
\alias{stat_density_watercolor}
\title{Bootstrapped Confidence Envelope for Density}
\arguments{
\item{data}{The data to be displayed in this layer.}

\item{mapping}{Default list of aesthetic mappings to use for plot.}

\item{k}{Number of bootstrapped smoothers. The default is 1000.}

\item{color}{Color for the bootstrapped densities that make up the confidence envelope. The
default is \code{color="#1D91C0"}}

\item{alpha}{Transparency level for the paths that make up the bootstrapped fitted lines. This may
need to be adjusted if the argument \verb{k=} is changed. The default value is \code{alpha=0.03}.}

\item{model}{The model to bootstrap from. The default is \code{model="none"} which bootstraps from the data.
Using \code{model="normal"} draws repeated samples from a normal distribution with parameters based on the data.}
}
\description{
This function creates a confidence envelope for the empirical density by bootstrapping from the data.
Transparency is used to indicate the level of uncertainty.
}
\section{Aesthetics}{

These stat uses  \code{geom_line()} so support the same aesthetics: \code{alpha}, \code{colour}, \code{group},
\code{linetype} and \code{size}.
}

