% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_residual_confidence.R
\name{stat_residual_confidence}
\alias{stat_residual_confidence}
\title{Confidence Envelope for Residuals}
\arguments{
\item{fill}{Fill color for the confidence envelope. The default is \code{fill="skyblue"}}

\item{model}{Name of fitted lm model.}
}
\description{
This function creates a normal theory based confidence envelope for the model residuals.
}
\section{Aesthetics}{

These stat uses  \code{geom_ribbon()} so support the same aesthetics: \code{alpha}, \code{colour}, \code{fill},
\code{linetype}, \code{group}, and \code{size}.
}

