% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_density_confidence.R
\docType{data}
\name{StatDensityConfidence}
\alias{StatDensityConfidence}
\title{Density with Confidence Envelope}
\format{An object of class \code{StatDensityConfidence} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.}
\usage{
StatDensityConfidence
}
\arguments{
\item{h}{A normal kernel function is used and `h` is its standard deviation. If this parameter is
omitted, a normal optimal smoothing parameter is used.}

\item{fill}{Fill color for the confidence envelope. The default is `color="skyblue"`}

\item{fade}{Should the confidence enevelope have more transparency for values farther from the
empirical density? The default is `fade=FALSE`.}

\item{model}{The model to bootstrap from. The default is `model="none"` which bootstraps from the data.
Using `model="normal"` draws repeated samples from a normal distribution with parameters based on the
ML estimates from the data.}
}
\description{
This function creates a normal theory based confidence envelope for the empirical density.
Transparency is used to indicate the level of uncertainty.
}
\keyword{datasets}
