% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatResidualConfidence.R
\docType{data}
\name{StatResidualConfidence}
\alias{StatResidualConfidence}
\title{Confidence Envelope for Residuals Helper Function}
\format{
An object of class \code{StatResidualConfidence} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.
}
\arguments{
\item{model}{Name of fitted lm model.}

\item{type}{Type of residual, either raw or standardized. The default is \code{type="standardized"}}

\item{fill}{Fill color for the confidence envelope. The default is \code{fill="skyblue"}}
}
\description{
This function creates a normal theory based confidence envelope for the model residuals.
}
\keyword{datasets}
