% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_watercolor_density.R
\docType{data}
\name{StatWatercolorDensity}
\alias{StatWatercolorDensity}
\title{Watercolor Bootstrapped Density}
\format{An object of class \code{StatWatercolorDensity} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 4.}
\usage{
StatWatercolorDensity
}
\arguments{
\item{k}{Number of bootstrapped smoothers. The default is 1000.}

\item{color}{Color for the bootstrapped fitted lines that make up the confidence envelope. The
default is `color="#1D91C0"`}

\item{alpha}{Transparency level for the paths that make up the bootstrapped fitted lines. This may
need to be adjusted if the argument `k=` is changed. The default value is `alpha=0.03`.}

\item{model}{The model to bootstrap from. The default is `model="none"` which bootstraps from the data.
Using `model="normal"` draws repeated samples from a normal distribution with parameters based on the
ML estimates from the data.}
}
\description{
This function creates a confidence envelope for the empirical density by bootstrapping from the data.
Transparency is used to indicate the level of uncertainty.
}
\keyword{datasets}
