% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_watercolor_smooth2.R
\docType{data}
\name{StatWatercolorSmoother2}
\alias{StatWatercolorSmoother2}
\title{Watercolor Bootstrapped Smoother}
\format{An object of class \code{StatWatercolorSmoother} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
StatWatercolorSmoother2
}
\arguments{
\item{k}{Number of bootstrapped smoothers. The default is 700.}

\item{color}{Color for the bootstrapped fitted lines that make up the confidence envelope. The
default is `color="#1D91C0"`}

\item{alpha}{Transparency level for the paths that make up the bootstrapped fitted lines. This may
need to be adjusted if the argument `k=` is changed. The default value is `alpha=0.06`.}

\item{method}{Smoothing method (function) to use, accepts either a character vector,
  e.g. `"auto"`, `"lm"`, `"glm"`, `"gam"`, `"loess"` or a function, e.g.
  `MASS::rlm` or `mgcv::gam`, `stats::lm`, or `stats::loess`.

  For `method = "auto"` the smoothing method is chosen based on the
  size of the largest group (across all panels). [stats::loess()] is
  used for less than 1,000 observations; otherwise [mgcv::gam()] is
  used with `formula = y ~ s(x, bs = "cs")`. Somewhat anecdotally,
  `loess` gives a better appearance, but is \eqn{O(N^{2})}{O(N^2)} in memory,
  so does not work for larger datasets.

  If you have fewer than 1,000 observations but want to use the same `gam()`
  model that `method = "auto"` would use, then set
  `method = "gam", formula = y ~ s(x, bs = "cs")`.}

\item{formula}{Formula to use in smoothing function. The default is `formula = y ~ x`,}

\item{fullrange}{Should the fit span the full range of the plot, or just
the data? The default is `fullrange=FALSE`.}

\item{span}{Controls the amount of smoothing for the default loess smoother.
Smaller numbers produce wigglier lines, larger numbers produce smoother
lines.}

\item{n}{Number of points at which to evaluate smoother.}

\item{method.args}{List of additional arguments passed on to the modelling
function defined by `method`.}
}
\description{
This function creates a visually regression smoother, also called a watercolor plot. The initial idea
came from Solomon Hsiang.
}
\keyword{datasets}
