% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_anova.R
\name{tidy_anova}
\alias{tidy_anova}
\title{Tidy a(n) ANOVA table}
\usage{
tidy_anova(x, model = FALSE)
}
\arguments{
\item{x}{an object containing the results returned by a model fitted with lm().}

\item{model}{Logical indicating whether the model-level partitioning (TRUE) or term-level partitioning (FALSE; default) should be used}
}
\value{
A data.frame with columns for the source of variation, degrees of freedom, Sum of Squares, Mean Squares, F-statistics, and p-values
}
\description{
Tidy a(n) ANOVA table
}
\examples{
lm.1 = lm(mpg ~ 1 + hp + wt, data = mtcars)
tidy_anova(lm.1, model = TRUE)

}
