% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watercolor_smoother.R
\name{watercolor_smoother}
\alias{watercolor_smoother}
\title{Watercolor Smoother}
\usage{
watercolor_smoother(data, x, y, k = 1000, method = "loess",
  color = "#1D91C0", alpha = 0.03, emp_line_color = "black",
  emp_line_size = 0.5, ...)
}
\arguments{
\item{data}{Dataframe}

\item{x}{A column name in the dataframe. This should be unquoted.}

\item{y}{A column name in the dataframe. This should be unquoted.}

\item{k}{Number of bootstrapped smoothers. The default is 1000.}

\item{method}{Fitting function for the bootstrapped fitted lines. The default is `method="loess"`.
The underlying syntax uses `stat_smooth()` so any method that works with that function should work.}

\item{color}{Color for the bootstrapped fitted lines that make up the confidence envelope. The
default is `color="#1D91C0"`}

\item{alpha}{Transparency level for the paths that make up the bootstrapped fitted lines. This may
need to be adjusted if the argument `k=` is changed. The default value is `alpha=0.03`.}

\item{emp_line_color}{Color of the fitted line for the empirical data.
The default is `emp_line_color="black"`.}

\item{emp_line_size}{Size of the fitted line for the empirical data. The default value is
`emp_line_size=0.5`.}

\item{...:}{Further parameters passed to the fitting function, in the case of loess, for example,
"span = .9", or "family = 'symmetric'"}
}
\value{
A ggplot object giving the plot of the bootstrapped fitted lines and the fitted line for
  the empirical data
}
\description{
This function creates a visually regression smoother, also called a watercolor plot. The initial idea
came from Solomon Hsiang.
}
