% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_results.R
\name{z_results}
\alias{z_results}
\title{Print z-Test Results}
\usage{
z_results(x)
}
\arguments{
\item{x}{prop.test() or prop_test() object}
}
\value{
Printed output of the z-test results including test type, hypotheses, z-value, and p-value.
}
\description{
Print z-Test Results
}
\examples{
my_z = prop_test(x = faithful$eruptions > 3, alternative = "less")
z_results(my_z)

}
