% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/educate-package.R
\docType{package}
\name{educate-package}
\alias{educate}
\alias{educate-package}
\title{educate: Miscellaneous R Functions for Educational Statistics}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

These are miscellaneous functions for carrying out statistical work in education.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zief0002/educate}
  \item Report bugs at \url{https://github.com/zief0002/educate/issues}
}

}
\author{
\strong{Maintainer}: Andrew Zieffler \email{zief0002@umn.edu} (\href{https://orcid.org/0000-0002-1035-0341}{ORCID}) [copyright holder]

}
\keyword{internal}
