% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_density_confidence.R,
%   R/stat_watercolor_density.R, R/stat_watercolor_smooth.R
\name{stat_density_confidence}
\alias{stat_density_confidence}
\alias{stat_watercolor_density}
\alias{stat_watercolor_smooth}
\title{Stat Method for Density with Confidence Envelope}
\arguments{
\item{h}{A normal kernel function is used and `h` is its standard deviation. If this parameter is
omitted, a normal optimal smoothing parameter is used.}

\item{fill}{Fill color for the confidence envelope. The default is `color="skyblue"`}

\item{fade}{Should the confidence enevelope have more transparency for values farther from the
empirical density? The default is `fade=FALSE`.}

\item{model}{The model to bootstrap from. The default is `model="none"` which bootstraps from the data.
Using `model="normal"` draws repeated samples from a normal distribution with parameters based on the
ML estimates from the data.}

\item{k}{Number of bootstrapped smoothers. The default is 1000.}

\item{alpha}{Transparency level for the paths that make up the bootstrapped fitted lines. This may
need to be adjusted if the argument `k=` is changed. The default value is `alpha=0.03`.}

\item{color}{Color for the bootstrapped fitted lines that make up the confidence envelope. The
default is `color="#1D91C0"`}

\item{formula}{Formula to use in smoothing function. The default is `formula = y ~ x`,}

\item{method}{Smoothing method (function) to use, accepts either a character vector,
  e.g. `"auto"`, `"lm"`, `"glm"`, `"gam"`, `"loess"` or a function, e.g.
  `MASS::rlm` or `mgcv::gam`, `stats::lm`, or `stats::loess`.

  For `method = "auto"` the smoothing method is chosen based on the
  size of the largest group (across all panels). [stats::loess()] is
  used for less than 1,000 observations; otherwise [mgcv::gam()] is
  used with `formula = y ~ s(x, bs = "cs")`. Somewhat anecdotally,
  `loess` gives a better appearance, but is \eqn{O(N^{2})}{O(N^2)} in memory,
  so does not work for larger datasets.

  If you have fewer than 1,000 observations but want to use the same `gam()`
  model that `method = "auto"` would use, then set
  `method = "gam", formula = y ~ s(x, bs = "cs")`.}

\item{n}{Number of points at which to evaluate smoother.}

\item{span}{Controls the amount of smoothing for the default loess smoother.
Smaller numbers produce wigglier lines, larger numbers produce smoother
lines.}

\item{fullrange}{Should the fit span the full range of the plot, or just
the data? The default is `fullrange=FALSE`.}

\item{method.args}{List of additional arguments passed on to the modelling
function defined by `method`.}
}
\description{
This function creates a confidence enevelope for the empirical density. The idea came from
functions written by Bowman and Azzalini.

This function creates a visually regression smoother, also called a watercolor plot. The initial idea
came from Solomon Hsiang.

This function creates a visually regression smoother, also called a watercolor plot. The initial idea
came from Solomon Hsiang.
}
\section{Aesthetics}{

These stat uses  `geom_ribbon()` so support the
same aesthetics: `alpha`, `colour`, `linetype` and
`size`.


These stat uses  `geom_line()` so support the
same aesthetics: `alpha`, `colour`, `linetype` and
`size`.


These stat uses  `geom_line()` so support the
same aesthetics: `alpha`, `colour`, `linetype` and
`size`.
}

