% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watercolor_density.R
\name{watercolor_density}
\alias{watercolor_density}
\title{Density plot with Bootstrapped Confidence Envelope}
\usage{
watercolor_density(data, x, k = 1000, boot_envelope = TRUE,
  model = "none", rug_color = "black", envelope_fill = "#7FCDBB",
  alpha = 0.03, emp_dens_color = "black", emp_dens_size = 0.5,
  multicolor = FALSE, palette = RColorBrewer::brewer.pal(9, "YlGnBu"),
  ...)
}
\arguments{
\item{data}{Dataframe}

\item{x}{Column name (unquoted)}

\item{k}{Number of bootstrap samples}

\item{boot_envelope}{Logical value to indicate whether the bootstrap confidence envelope should be displayed}

\item{model}{Character string indicating the type of bootstrapped confidence envelope to display;}

\item{rug_color}{Color for the rug plot}

\item{envelope_fill}{Fill color for the bootstrapped confidence envelope}

\item{alpha}{Transparency level for the paths that make up the bootstrapped confidence envelope}

\item{emp_dens_color}{Color of the line for the density plot of the empirical data}

\item{emp_dens_size}{Size of the line for the density plot of the empirical data}

\item{multicolor}{Should the confidence enevelope be multicolored or monocolored}

\item{palette}{Color palette used if multicolor=TRUE}
}
\value{
A ggplot object giving the density plot
}
\description{
This function creates a visually weighted coefficient plot
}
